+++ packages/eslint-plugin-best-practices/CHANGELOG.md
@@ -2,3 +2,4 @@
 
 ## 0.2.8
 - Update ESLint 7.x
+- Overrides `**/__tests__/*.{j,t}s?(x)` and `**/tests/*.{j,t}s?(x)` env config to `jest: true`